Refactoring: Abstract notification token manage service using interface
